Added a function that dynamically creates fixtures with a certain scope
The trick to making this work is that pytest finds fixtures by scanning
the namespace of the module with tests for valid fixtures. Basically,
the fixture decorator generates a valid fixture but doesn't register it
anywhere. The fixture just has to be present in the namespace of the
testing functions that require it and than it's picked up automatically.

So I added a function to the plugin that returns a valid fixture
function. To use it you just need to bind it to a variable in your
module. Invocation is thus like:
-----------
my_fancy_fixture = generate_scoped_network_info_fixture('class')

def test_my_fixture(my_fancy_fixture):
    assert my_fancy_fixture