Changed the dynamically scoped fixture to return containers instead of network_info
The network_info does not quite contain everything I need so I want to
have access to the containers instead. To still get the convenience of
the network_info wrapper I added those to the containers returned by the
fixture.

In addition the containers are return in a dict instead of a list.