Changed the api port back to a static configuration
tox does not properly expose the port when dynamically allocated in the
circleci VM. If I manually ssh into the VM and I spin up the network it works
just fine. It's really the tox layer causing the issue. Tox does work for
statically allocated ports so I'm reverting back to that.

If I run tox locally with the dynamically allocated ports it works like a charm
so I don't know what's wrong. Maybe something with tox-pyenv?